Consolidate FlashVSR uplift gRPC modules
